Issues fixed:
1 Logic from one OrderController divided to 4 controllers: viewOrder, viewAllOrders, deleteOrder and CompleteOrder
