+++ src/main/java/mate/academy/internetshop/controller/DeleteOrderController.java
@@ -1,0 +1,32 @@
+package mate.academy.internetshop.controller;
+
+import java.io.IOException;
+import java.util.List;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import mate.academy.internetshop.lib.Injector;
+import mate.academy.internetshop.model.Order;
+import mate.academy.internetshop.model.User;
+import mate.academy.internetshop.service.OrderService;
+import mate.academy.internetshop.service.UserService;
+
+public class DeleteOrderController extends HttpServlet {
+    private static Long USER_ID = 1L;
+    private static final Injector INJECTOR = Injector.getInstance("mate.academy.internetshop");
+    private UserService userService
+            = (UserService) INJECTOR.getInstance(UserService.class);
+    private OrderService orderService
+            = (OrderService) INJECTOR.getInstance(OrderService.class);
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
+            throws ServletException, IOException {
+        User user = userService.get(USER_ID);
+        orderService.delete(Long.valueOf(req.getParameter("orderId")));
+        List<Order> userOrders = orderService.getUserOrders(user);
+        req.setAttribute("userOrders", userOrders);
+        req.getRequestDispatcher("/WEB-INF/views/orders/allOrders.jsp").forward(req, resp);
+    }
+}
