+++ src/main/java/mate/academy/internetshop/controller/IndexController.java
@@ -1,4 +1,4 @@
-package mate.academy.internetshop.controllers;
+package mate.academy.internetshop.controller;
 
 import java.io.IOException;
 import java.time.LocalTime;
