+++ src/main/java/mate/academy/internetshop/controller/ViewOrderController.java
@@ -1,0 +1,25 @@
+package mate.academy.internetshop.controller;
+
+import java.io.IOException;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import mate.academy.internetshop.lib.Injector;
+import mate.academy.internetshop.model.Order;
+import mate.academy.internetshop.service.OrderService;
+
+public class ViewOrderController extends HttpServlet {
+    private static final Injector INJECTOR = Injector.getInstance("mate.academy.internetshop");
+    private OrderService orderService
+            = (OrderService) INJECTOR.getInstance(OrderService.class);
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
+            throws ServletException, IOException {
+        Order order = orderService.get(Long.valueOf(req.getParameter("orderId")));
+        req.setAttribute("products", order.getProducts());
+        req.getRequestDispatcher("/WEB-INF/views/orders/order.jsp").forward(req, resp);
+
+    }
+}
